gamestate: Rename GameSimulation file.
