+++ libopenage/gamestate/simulation.h
