+++ libopenage/main.cpp
@@ -7,7 +7,7 @@
 
 #include "cvar/cvar.h"
 #include "event/time_loop.h"
-#include "gamestate/engine.h"
+#include "gamestate/simulation.h"
 #include "log/log.h"
 #include "presenter/presenter.h"
 #include "util/timer.h"
